moved BG calc function to homepage_state
