+++ lib/pages/home_page.dart
@@ -1,4 +1,5 @@
 import 'package:calcal/reuse.dart';
+import 'package:calcal/states/homepage_state.dart';
 import 'package:calcal/values.dart';
 import 'package:flutter/material.dart';
 
@@ -9,16 +10,6 @@ class home_page extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    Color calculateBackgroundColor({required double value}) {
-      if (value > 0.60) {
-        return Colors.red;
-      } else if (value > 0.30) {
-        return Colors.orange;
-      } else {
-        return Colors.green;
-      }
-    }
-
     return SafeArea(
       child: Stack(
         children: [
@@ -47,7 +38,7 @@ class home_page extends StatelessWidget {
                         value: 0.3,
                         backgroundColor: palette.progressbgColor,
                         valueColor: AlwaysStoppedAnimation<Color>(
-                          calculateBackgroundColor(value: 0.3),
+                          homepage_state().calculateBackgroundColor(value: 0.3),
                         ),
                         strokeWidth: 8,
                       ),
